requirements.txt
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.12.0
pillow>=9.0.0
jupyter>=1.0.0
opencv-python>=4.6.0
tensorflow>=2.10.0  # For CNN models (retinal/product images)
torch>=1.13.0  # Alternative to TensorFlow
streamlit>=1.35.0
scikeras>=0.13.0
pyyaml>=6.0.1
gdown